% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrationPlots.R
\name{consolidateConcentrationData}
\alias{consolidateConcentrationData}
\title{Consolidate concentration data}
\usage{
consolidateConcentrationData(directory = "")
}
\arguments{
\item{directory}{The directory from which to load the data.
Default: c("EGF", "oxygen")}
}
\value{
A single list of data frames, one for each element to plot.
}
\description{
Consolidate concentration data
}
