% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatePlots.R
\name{consolidateFateData}
\alias{consolidateFateData}
\title{Consolidate fate data}
\usage{
consolidateFateData(directory = "")
}
\arguments{
\item{directory}{The directory from which to load the data.}
}
\value{
A single dataframe in long format.
}
\description{
Consolidate fate data
}
