% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probPlots.R
\name{consolidatePbnSimulationData}
\alias{consolidatePbnSimulationData}
\title{Consolidate PBN simulation data}
\usage{
consolidatePbnSimulationData(
  directory = "",
  experiment.tag = "default",
  probabilities,
  n.files
)
}
\arguments{
\item{directory}{The directory from which to load the data.}

\item{experiment.tag}{The experiment name or tag.}

\item{probabilities}{A vector of the different probabilities used for each simulation.}

\item{n.files}{The number of files each simulation is divided across.}
}
\value{
A single list of data frames, one for each simulation.
}
\description{
Consolidate PBN simulation data
}
