% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{loadAndBind}
\alias{loadAndBind}
\title{Load raw microCJ data and bind it into a single data frame}
\usage{
loadAndBind(directory, not.numeric = c("cell_id", "subpopulation"))
}
\arguments{
\item{directory}{the directory where the data is located}
}
\value{
a data frame with the integrated data
}
\description{
Load raw microCJ data and bind it into a single data frame
}
