% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPlots.R
\name{loadTimePointData}
\alias{loadTimePointData}
\title{Load timepoint data}
\usage{
loadTimePointData(timepoints, dir, mode, suffix)
}
\arguments{
\item{timepoints}{A vector with the timepoints involved}

\item{dir}{The directory from which to load the data}

\item{mode}{The type of timepoint stat}

\item{suffix}{The suffix of the filename, such as "_1.csv"}
}
\value{
A dataframe with the timepoint data
}
\description{
Load timepoint data
}
