% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcjPlotter-package.R
\docType{package}
\name{mcjPlotter-package}
\alias{mcjPlotter}
\alias{mcjPlotter-package}
\title{mcjPlotter: Plot MicroCJ Simulation Data}
\description{
Helps plotting MicroCJ simulation data and cleaning the output files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CBigOxf/mcjPlotter}
  \item Report bugs at \url{https://github.com/CBigOxf/mcjPlotter/issues}
}

}
\author{
\strong{Maintainer}: Pedro Victori \email{pedroluisvictori@gmail.com} (\href{https://orcid.org/0000-0001-6321-5584}{ORCID})

}
\keyword{internal}
