% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPlots.R
\name{plotCellData}
\alias{plotCellData}
\title{Plot timepoint data}
\usage{
plotCellData(
  data,
  by,
  r = 10,
  discr = F,
  directory = "",
  filename = "cellPlot",
  width = 10,
  height = 10,
  viridisOption = "D",
  facet = T
)
}
\arguments{
\item{data}{The data to be plotted}

\item{by}{The name of the column represented in the cell color}
}
\value{
A ggplot2 object
}
\description{
Plot timepoint data
}
