% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrationPlots.R
\name{plotConcentration}
\alias{plotConcentration}
\title{Plot concentrations of elements of interest}
\usage{
plotConcentration(
  data,
  directory = "",
  filename = "concentrationPlots",
  to.plot,
  viridis_palette = "magma",
  fill_direction = 1,
  plot.title = " concentration",
  in.width = 9,
  in.height = 9
)
}
\arguments{
\item{data}{The input data,
generated by \code{\link{consolidateConcentrationData}}.}

\item{directory}{The directory to which the plot should be saved.}

\item{to.plot}{Element to plot. Default: EGF.}

\item{plot.title}{The plot title.}

\item{in.width}{Width in inches.}

\item{in.height}{Height in inches.}

\item{plot.subtitle}{The plot subtitle.}
}
\value{
A list of with heatmap plots of concentrations per i,j.
}
\description{
Plot concentrations of elements of interest
}
