% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{plotCounts}
\alias{plotCounts}
\title{Plot standard microCJ data}
\usage{
plotCounts(data, timesteps = 3000)
}
\arguments{
\item{data}{microCJ data}

\item{timesteps}{how many steps in the data}
}
\value{
A ggplot2 object
}
\description{
Plot standard microCJ data
}
