% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatePlots.R
\name{plotFate}
\alias{plotFate}
\title{Plot summary of fate of mutants of interest in stacked area}
\usage{
plotFate(
  data,
  directory = "",
  filename = "nodeStatusPlots",
  apoptosis.cumulative = TRUE,
  plot.title = "Number of cells reaching each fate",
  plot.subtitle = "Per timepoint and mutant",
  in.width = 9,
  in.height = 7
)
}
\arguments{
\item{data}{The input data,
generated by \code{\link{consolidateFateData}}.}

\item{directory}{The directory to which the plot should be saved.}

\item{plot.title}{The plot title.}

\item{plot.subtitle}{The plot subtitle.}

\item{in.width}{Width in inches.}

\item{in.height}{Height in inches.}
}
\value{
A list of stacked area plots of number of cells reaching
fate decisions per timepoint per mutant.
}
\description{
Plot summary of fate of mutants of interest in stacked area
}
