% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodePlots.R
\name{plotNodes}
\alias{plotNodes}
\title{Plot node data}
\usage{
plotNodes(
  data,
  timepoints = c(),
  filename,
  width = 18,
  height = 10,
  title = "Node status, average of all cells in each subpopulation",
  subtitle = "100 replicates per time point"
)
}
\arguments{
\item{data}{processed data}

\item{timepoints}{time points to include}

\item{filename}{file name to save the plot}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}
}
\value{
the plot
}
\description{
Plot node data
}
