% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probPlots.R
\name{plotPbnSimulation}
\alias{plotPbnSimulation}
\title{Plot consolidated PBN simulation data}
\usage{
plotPbnSimulation(
  data,
  directory = "",
  plot.title = "PBN simulation",
  plot.subtitle = "Average of all replicates and standard error",
  timesteps,
  pop.labels
)
}
\arguments{
\item{data}{The input data, generated by \code{\link{consolidatePbnSimulationData}}.}

\item{directory}{The directory to which the plot should be saved.}

\item{plot.title}{The plot title.}

\item{plot.subtitle}{The plot subtitle.}

\item{timesteps}{Number of time steps to plot.}

\item{pop.labels}{A vector of each population's label.}
}
\value{
A list of all plots.
}
\description{
Plot consolidated PBN simulation data
}
