% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodePlots.R
\name{processNodeData}
\alias{processNodeData}
\title{Process node data}
\usage{
processNodeData(data, filename)
}
\arguments{
\item{data}{the data to be processed}

\item{filename}{the file name to save processed data.}
}
\value{
the processed data
}
\description{
Process node data
}
