% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrationPlots.R
\name{consolidateConcentrationData}
\alias{consolidateConcentrationData}
\title{Consolidate concentration data}
\usage{
consolidateConcentrationData(directory = "", to.plot = c("EGF", "oxygen"))
}
\arguments{
\item{directory}{The directory from which to load the data.}

\item{to.plot}{Elements of interest to plot later. Default: c("EGF", "oxygen")}
}
\value{
A single list of data frames, one for each element to plot.
}
\description{
Consolidate concentration data
}
