% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatePlots.R
\name{consolidateFateData}
\alias{consolidateFateData}
\title{Consolidate fate data}
\usage{
consolidateFateData(
  directory = "",
  mutants.to.plot = c("EGFR+", "p53-PTEN-", "EGFR+p53-PTEN-", "WT")
)
}
\arguments{
\item{directory}{The directory from which to load the data.}

\item{mutants.to.plot}{Mutants to plot in separate figures later.}
}
\value{
A single list of data frames, one for each element to plot.
}
\description{
Consolidate fate data
}
