% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrationPlots.R
\name{plotConcentration}
\alias{plotConcentration}
\title{Plot concentrations of elements of interest}
\usage{
plotConcentration(
  data,
  directory = "",
  filename = "concentrationPlots",
  plot.title = "EGF and oxygen concentration",
  plot.subtitle = "Per i and j coordinates",
  in.width = 9,
  in.height = 9
)
}
\arguments{
\item{data}{The input data,
generated by \code{\link{consolidateConcentrationData}}.}

\item{directory}{The directory to which the plot should be saved.}

\item{plot.title}{The plot title.}

\item{plot.subtitle}{The plot subtitle.}

\item{in.width}{Width in inches.}

\item{in.height}{Height in inches.}
}
\value{
A list of with heatmap plots of concentrations per i,j.
}
\description{
Plot concentrations of elements of interest
}
