% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodePlots.R
\name{plotNodes}
\alias{plotNodes}
\title{Plot node data}
\usage{
plotNodes(data, timepoints = c(), filename, width = 18, height = 10)
}
\arguments{
\item{data}{processed data}

\item{timepoints}{time points to include}

\item{filename}{file name to save the plot}

\item{width}{width of the plot}

\item{height}{height of the plot}
}
\value{
the plot
}
\description{
Plot node data
}
